Moved ConfirmOrder function and logic from CartController to OrderController in webapp
