+++ Labb1/Controllers/CartController.cs
@@ -104,73 +104,75 @@ namespace Labb1.Controllers
             }
         }
 
-        [HttpPost]
-        public async Task<IActionResult> ConfirmOrder([Bind("TotalPrice, CartItems")]CartViewModel cartViewModel)
-        {
-            // Error handling. 
-            if (await SessionHelper.GetObjectFromJson<List<CartItem>>(HttpContext.Session, "cart") != null)
-            {
-                // get the cart from session
-                var cart = await GetCart();
+        //[HttpPost]
+        //public async Task<IActionResult> ConfirmOrder([Bind("TotalPrice, CartItems")]CartViewModel cartViewModel)
+        //{
+        //    // Error handling. 
+        //    if (await SessionHelper.GetObjectFromJson<List<CartItem>>(HttpContext.Session, "cart") != null)
+        //    {
+        //        // get the cart from session
+        //        var cart = await GetCart();
 
-                var user = await _userManager.GetUserAsync(HttpContext.User);
-                if (String.IsNullOrEmpty(user.FirstName) || String.IsNullOrEmpty(user.FirstName) ||
-                    String.IsNullOrEmpty(user.StreetAddress) || String.IsNullOrEmpty(user.ZipCode) ||
-                    String.IsNullOrEmpty(user.City))
-                {
-                    TempData["MessageAddressInfo"] = "Vänligen fyll i ditt namn och adress innan du kan beställa";
-                    return LocalRedirect("/Identity/Account/Manage");
-                }
-                else
-                {
-                    // Add products in order for each product in cart
-                    var orderProducts = cart.Select(x => new OrderProduct
-                {
-                    Product = x.Product,
-                    Quantity = x.Quantity
-                }).ToList();
+        //        var user = await _userManager.GetUserAsync(HttpContext.User);
+        //        if (String.IsNullOrEmpty(user.FirstName) || String.IsNullOrEmpty(user.FirstName) ||
+        //            String.IsNullOrEmpty(user.StreetAddress) || String.IsNullOrEmpty(user.ZipCode) ||
+        //            String.IsNullOrEmpty(user.City))
+        //        {
+        //            TempData["MessageAddressInfo"] = "Vänligen fyll i ditt namn och adress innan du kan beställa";
+        //            return LocalRedirect("/Identity/Account/Manage");
+        //        }
+        //        else
+        //        {
+        //            // Add products in order for each product in cart
+        //            var orderProducts = cart.Select(x => new OrderProduct
+        //        {
+        //            Product = x.Product,
+        //            Quantity = x.Quantity
+        //        }).ToList();
 
-                    // Remove all products in cart where quantity is 0
-                    var itemsToRemove = from items in orderProducts
-                                        where items.Quantity == 0
-                                        select items;
-                    foreach (var item in itemsToRemove.ToList())
-                    {
-                        orderProducts.Remove(item); 
-                    }
-                    // Prevent user to confirm order without products in cart
-                    if (orderProducts.Count == 0)
-                    {
-                        TempData["MessageCartInfo"] = "Du kan inte lägga en order, din kundkorg är tom.";
-                        return RedirectToAction("Index");
-                    }
-
-
-                    OrderViewModel orderViewModel = new OrderViewModel();
-                    Order order = new Order
-                    {
-                        OrderDate = DateTime.Now,
-                        UserId = Guid.Parse(_userManager.GetUserId(User))
-                    };
+        //            // Remove all products in cart where quantity is 0
+        //            var itemsToRemove = from items in orderProducts
+        //                                where items.Quantity == 0
+        //                                select items;
+        //            foreach (var item in itemsToRemove.ToList())
+        //            {
+        //                orderProducts.Remove(item); 
+        //            }
+        //            // Prevent user to confirm order without products in cart
+        //            if (orderProducts.Count == 0)
+        //            {
+        //                TempData["MessageCartInfo"] = "Du kan inte lägga en order, din kundkorg är tom.";
+        //                return RedirectToAction("Index");
+        //            }
 
 
-                orderViewModel.Order = order;
-                orderViewModel.User = user;
-                orderViewModel.Order.OrderProducts = orderProducts;
-                orderViewModel.Order.TotalPrice = cartViewModel.TotalPrice;
+        //            OrderViewModel orderViewModel = new OrderViewModel();
+        //            Order order = new Order
+        //            {
+        //                OrderDate = DateTime.Now,
+        //                UserId = Guid.Parse(_userManager.GetUserId(User))
+        //            };
+
+
+        //        orderViewModel.Order = order;
+        //        orderViewModel.User = user;
+        //        orderViewModel.Order.OrderProducts = orderProducts;
+        //        orderViewModel.Order.TotalPrice = cartViewModel.TotalPrice;
 
                   
-                // Clear cart session
-                HttpContext.Session.Remove("cart");
+        //        // Clear cart session
+        //        HttpContext.Session.Remove("cart");
 
-                return View("OrderConfirmed", orderViewModel);
-                }
-            }
-            else
-            {
-                return NotFound();
-            }
-        }
+        //            //return View("OrderConfirmed", orderViewModel);
+        //            TempData["Order"] = orderViewModel;
+        //            return RedirectToAction("Index", "Order");
+        //        }
+        //    }
+        //    else
+        //    {
+        //        return NotFound();
+        //    }
+        //}
         public async Task<CartViewModel> ShowCart()
         {
             var cart = await GetCart();
