+++ Labb1/Views/Cart/OrderConfirmed.cshtml
@@ -1,4 +1,4 @@
-﻿@model Labb1.ViewModels.OrderViewModel
+﻿@*@model Labb1.ViewModels.OrderViewModel
 
 @{
     ViewData["Title"] = "OrderConfirmed";
@@ -64,4 +64,4 @@
             </div>
         </div>
     </div>
-</div>
+</div>*@
